// src/types/biometricsTypes.ts
// Defines fingerprint data structure
export type FingerprintData = {
    left_index: string | null;
    left_middle: string | null;
    left_ring: string | null;
    left_pinky: string | null;
    left_thumb: string | null;
    right_index: string | null;
    right_middle: string | null;
    right_ring: string | null;
    right_pinky: string | null;
    right_thumb: string | null;
    profile_photo: string | null;   
};